+++ docs/changelog.rst
@@ -7,11 +7,14 @@ Changelog
 Version 0.6.0 - Unreleased
 **************************
 
+Adds **Foundation for site 6** support, version 5 is still available.
+
 * Copied 'foundation-6' templates from @flesser branch ``foundation-6``;
 * Fixed demo views and forms so they can switch between template packs;
 * Fixed layout elements so their template does not include ``TEMPLATE_PACK`` anymore in class defintions;
 * Fixed switches;
 * Fixed button group;
+* ``.help-text`` is allways a ``<p>`` in Foundation6 (does not have real meaning in Foundation5);
 
 Version 0.5.5 - 2017/02/01
 **************************