+++ docs/changelog.rst
@@ -7,7 +7,7 @@ Changelog
 Version 0.6.0 - Unreleased
 **************************
 
-Adds **Foundation for site 6** support, version 5 support is still available for now.
+This release adds **Foundation for site version 6** support, version 5 support is still available for now.
 
 * Copied 'foundation-6' templates from @flesser branch ``foundation-6``;
 * Fixed demo views and forms so they can switch between template packs;
@@ -16,6 +16,7 @@ Adds **Foundation for site 6** support, version 5 support is still available for
 * Fixed button group for Foundation-6;
 * Fixed ``InlineJustifiedField`` for Foundation-6;
 * Fixed error messages for Foundation-6;
+* Fixed Accordion for Foundation-6;
 * Added ``base.Callout`` element;
 * ``.help-text`` is allways a ``<p>`` in Foundation6 (does not have real meaning in Foundation5);
 * Added ``crispy_forms_foundation.templatetags.crispy_forms_foundation_field`` to re-implement ``crispy_field`` filter so we can have the right input field error for Foundation-6;